organizations/requirements.txt
alembic==1.7.7
amqp==5.1.0
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dbfread==2.0.7
defusedxml==0.7.1
docxcompose==1.3.4
docxtpl==0.15.2
flake8==4.0.1
Flask==2.0.3
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fpdf==1.7.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
kombu==5.2.4
lxml==4.8.0
Mako==1.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
odfpy==1.4.1
packaging==21.3
Petrovich==2.0.1
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
pycdlib==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.1
python-docx==0.8.11
python-dotenv==0.20.0
pytils==0.4.1
pytz==2022.1
requests==2.27.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.32
tablib==3.2.0
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2024.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
